<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastSEAscore.R
\name{FastSEAscore}
\alias{FastSEAscore}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Fast calculate phenotypic set enrichment score.
}
\details{
FastSEAscore
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastSEAscore.R
\name{FastSEAscore}
\alias{FastSEAscore}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Fast calculate phenotypic set enrichment score.
}
\details{
FastSEAscore
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
>>>>>>> c43b0c8084a22940352538ed5e0734e5a168a738
