<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPackageLoaded.R
\name{isPackageLoaded}
\alias{isPackageLoaded}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Determine if the package is loaded, if no package is loaded.
}
\details{
isPackageLoaded
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPackageLoaded.R
\name{isPackageLoaded}
\alias{isPackageLoaded}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Determine if the package is loaded, if no package is loaded.
}
\details{
isPackageLoaded
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
>>>>>>> c43b0c8084a22940352538ed5e0734e5a168a738
