<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test_m.R
\name{ks_test_m}
\alias{ks_test_m}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Calculating subpathway Variation score.
}
\details{
ks_test_m
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test_m.R
\name{ks_test_m}
\alias{ks_test_m}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Calculating subpathway Variation score.
}
\details{
ks_test_m
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
>>>>>>> c43b0c8084a22940352538ed5e0734e5a168a738
