% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpwACmap.R
\name{plotSpwACmap}
\alias{plotSpwACmap}
\title{Plot subpathway activity change map}
\usage{
plotSpwACmap(inputdata, spwid = "")
}
\arguments{
\item{inputdata}{A list of result data generated by function `SubSEA` or `DCSA`.}

\item{spwid}{The subpathway id which the user wants to plot.}
}
\value{
a plot
}
\description{
Plot a box diagram and heat map of subpathway activity in each phenotype.
}
\details{
plotSpwACmap


Plot a box diagram of subpathway activity in each Phenotype and a heat map of the distribution of the phenotypic samples in the activity of the subpathways.
The subpathway activity change map includes subpathway active change box plot and subpathway active change. Each row in the heat map is all samples of a phenotype.
These samples are distributed in the subpathway high activity value area label is red, and the distribution in the low expression value area label is blue.
}
\examples{
# get the Subspwresult which is the result of SubSEA method.
Subspwresult<-get("Subspwresult")
# plot the subpathway 00120_9 in the SubSEA function result.
plotSpwACmap(Subspwresult,spwid="00120_9")
# get the DCspwresult which is the result of DCSA method.
DCspwresult<-get("DCspwresult")
# plot the subpathway 00982_2 in the DCSA function result.
plotSpwACmap(DCspwresult,spwid="00982_2")
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
