% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.R
\name{ssgsea}
\alias{ssgsea}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Single sample GSEA (ssGSEA) calculates a gene set enrichment score
per sample as the normalized difference in empirical cumulative distribution
functions of gene expression ranks inside and outside the gene set.
}
\details{
ssgsea
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
