% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rank_score.R
\name{compute_rank_score}
\alias{compute_rank_score}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Compute rank score.
}
\details{
compute_rank_score
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
