<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgenedensity.R
\name{getgenedensity}
\alias{getgenedensity}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Calculating the Kernel estimation of gene.
}
\details{
getgenedensity
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgenedensity.R
\name{getgenedensity}
\alias{getgenedensity}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Calculating the Kernel estimation of gene.
}
\details{
getgenedensity
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
>>>>>>> c43b0c8084a22940352538ed5e0734e5a168a738
