<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadClsFile.R
\name{ReadClsFile}
\alias{ReadClsFile}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
These are functions read sample label file (.cls format).
}
\details{
ReadClsFile
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadClsFile.R
\name{ReadClsFile}
\alias{ReadClsFile}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
These are functions read sample label file (.cls format).
}
\details{
ReadClsFile
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
>>>>>>> c43b0c8084a22940352538ed5e0734e5a168a738
