% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndWalk.R
\name{rndWalk}
\alias{rndWalk}
\title{psSubpathway internal functions}
\usage{
SubSEA
}
\description{
Calculating random walks.
}
\details{
rndWalk
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
