% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAscore.R
\name{SEAscore}
\alias{SEAscore}
\title{psSubpathway internal functions}
\usage{
plotSubSEScurve
}
\description{
Get subtype set enrichment score and sample locations, etc.
}
\details{
SEAscore
}
\author{
Xudong Han,
Junwei Han,
Qingfei Kong
}
